Полностью рабочий QR, только нужно Сделать нормальные текста
+ Начал работу над расписанием

+++ NS/Model NS/Bus.swift
@@ -2,10 +2,10 @@
 import SwiftUI
 
 struct Bus: Codable{
-    let id: Int64
-    let drive_id: Int64
+    let id: Int
+    let drive_id: Int
     let model: String
     let code: String
     let status: String
-    let number_of_sits: Int64
+    let number_of_sits: Int
 }
