+++ NS/Model NS/DailySchedule.swift
@@ -1,9 +1,10 @@
 import SwiftUI
 
 struct DailySchedule: Codable {
-    let id: Int64
+    let id: Int
     let date: Date
     let status: String
     let tickets: [Ticket]
     let trip: Trip
 }
+
