+++ NS/Model NS/Driver.swift
@@ -2,7 +2,7 @@ import SwiftUI
 
 struct Driver: Codable{
     let token: String?
-    let id: Int64
+    let id: Int
     let first_name: String
     let last_name: String
     let father_name: String
@@ -10,5 +10,5 @@ struct Driver: Codable{
     let phone_number: String
     let role: String
     let driver_id: String
-    let bus_id: Int64
+    let bus_id: Int
 }
